requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.20.16
botocore==1.23.16
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.8
Click==7.0
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
favicon==0.7.0
feedparser==6.0.8
filelock==3.4.0
fonttools==4.28.5
frozenlist==1.2.0
future==0.18.2
greenlet==1.1.2
httplib2==0.14.0
idna==2.8
jmespath==0.10.0
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.4
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
multidict==5.2.0
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.1
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
s3transfer==0.5.0
SecretStorage==2.3.1
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
systemd-python==234
tldextract==3.1.2
toml==0.10.2
typing-extensions==4.0.0
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.0
wadllib==1.3.3
websocket-client==1.2.3
websockets==10.1
xkit==0.0.0
yarl==1.7.2
